test(validate): organizes cyce tests around via and via-only
corrects an error in the JSON schema
