+++ src/schema/cruise-result.schema.json
@@ -670,7 +670,6 @@
       }
     },
     "MiniDependencyRestrictionType": {
-      "required": [],
       "oneOf": [
         { "$ref": "#/definitions/REAsStringsType" },
         {
