+++ test/validate/index.cycle-via.spec.mjs
@@ -198,193 +198,87 @@ describe("[I] validate/index dependency - cycle via - with group matching", () =
   });
 });
 
-describe("[I] validate/index dependency - cycle viaOnly", () => {
-  const lCycleViaRuleSet = parseRuleSet({
+describe("[I] validate/index dependency - cycle viaSomeNot (normalizes to via.pathNot) - with group matching", () => {
+  const lCycleButNotViaGroupMatchRuleSet = {
     forbidden: [
       {
-        from: {},
-        to: {
-          circular: true,
-          viaOnly: { path: "^tmp/ab\\.js$" },
-        },
+        name: "no-circular-dependency-of-modules",
+        from: { path: "^src/([^/]+)/.+" },
+        to: { viaSomeNot: "^src/$1/.+", circular: true },
       },
     ],
-  });
-  const lCycleViaNotTypeOnlyRuleSet = parseRuleSet({
-    forbidden: [
-      {
-        name: "no-runtime-cycles",
-        from: {},
-        severity: "error",
-        to: {
-          circular: true,
-          viaOnly: { dependencyTypesNot: ["type-only"] },
-        },
-      },
-    ],
-  });
-  it("a => ba => bb => bc => a doesn't get flagged when the cycle doesn't go via the viaOnly", () => {
-    deepEqual(
-      validate.dependency(
-        lCycleViaRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/ba.js",
-          circular: true,
-          cycle: ["tmp/ba.js", "tmp/bb.js", "tmp/bc.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
-        },
-      ),
-      {
-        valid: true,
-      },
-    );
-  });
+  };
 
-  it("a => aa => ab => ac => does not get flagged when only some of them are not in the viaOnly", () => {
+  it("flags when all of the cycle (except the root) is outside the group-matched viaSomeNot", () => {
     deepEqual(
       validate.dependency(
-        lCycleViaRuleSet,
-        { source: "tmp/a.js" },
+        parseRuleSet(lCycleButNotViaGroupMatchRuleSet),
+        { source: "src/module-a/a.js" },
         {
-          resolved: "tmp/aa.js",
+          resolved: "src/module-b/ba.js",
           circular: true,
-          cycle: ["tmp/aa.js", "tmp/ab.js", "tmp/ac.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
-        },
-      ),
-      {
-        valid: true,
-      },
-    );
-  });
-
-  it("a => ab a gets flagged because all of the via's in the cycle are in the viaOnly", () => {
-    deepEqual(
-      validate.dependency(
-        lCycleViaRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: ["tmp/ab.js", "tmp/a.js"].map(stringToCycleEntry),
-        },
-      ),
-      {
-        valid: true,
-      },
-    );
-  });
-
-  it("a => aa => ab => ac => a get flagged when all of them are in a viaOnly", () => {
-    const lRuleSet = parseRuleSet({
-      forbidden: [
-        {
-          from: {},
-          to: {
-            circular: true,
-            via: "^tmp/[^.]+\\.js$",
-          },
-        },
-      ],
-    });
-    deepEqual(
-      validate.dependency(
-        lRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: ["tmp/aa.js", "tmp/ab.js", "tmp/ac.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
+          cycle: [
+            "src/module-b/ba.js",
+            "src/module-b/bb.js",
+            "src/module-b/bc.js",
+            "src/module-a/a.js",
+          ].map(stringToCycleEntry),
         },
       ),
       {
         valid: false,
-        rules: [{ name: "unnamed", severity: "warn" }],
-      },
-    );
-  });
-  it("a => aa => ab => ac => a doesn't get flagged when one of the dependencyTypes is in a viaNot", () => {
-    deepEqual(
-      validate.dependency(
-        lCycleViaNotTypeOnlyRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: [
-            { name: "tmp/aa.js", dependencyTypes: ["import"] },
-            { name: "tmp/ab.js", dependencyTypes: ["import", "type-only"] },
-            { name: "tmp/ac.js", dependencyTypes: ["import"] },
-            { name: "tmp/a.js", dependencyTypes: ["import"] },
-          ],
-        },
-      ),
-      {
-        valid: true,
-      },
-    );
-  });
-
-  it("a => aa => ab => ac => a does get flagged when none of the dependencyTypes is in a viaNot", () => {
-    deepEqual(
-      validate.dependency(
-        lCycleViaNotTypeOnlyRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: [
-            { name: "tmp/aa.js", dependencyTypes: ["import"] },
-            { name: "tmp/ab.js", dependencyTypes: ["import"] },
-            { name: "tmp/ac.js", dependencyTypes: ["import"] },
-            { name: "tmp/a.js", dependencyTypes: ["import"] },
-          ],
-        },
-      ),
-      {
         rules: [
           {
-            name: "no-runtime-cycles",
-            severity: "error",
+            name: "no-circular-dependency-of-modules",
+            severity: "warn",
           },
         ],
+      },
+    );
+  });
+
+  it("flags when only one of the cycle is outside the group-matched viaNot", () => {
+    deepEqual(
+      validate.dependency(
+        parseRuleSet(lCycleButNotViaGroupMatchRuleSet),
+        { source: "src/module-a/a.js" },
+        {
+          resolved: "src/module-a/aa.js",
+          circular: true,
+          cycle: [
+            "src/module-a/aa.js",
+            "src/module-a/ab.js",
+            "src/module-b/bc.js",
+            "src/module-a/a.js",
+          ].map(stringToCycleEntry),
+        },
+      ),
+      {
         valid: false,
+        rules: [
+          {
+            name: "no-circular-dependency-of-modules",
+            severity: "warn",
+          },
+        ],
       },
     );
   });
-});
 
-describe("[I] validate/index dependency - cycle viaOnly - with group matching", () => {
-  const lCycleViaRuleSet = parseRuleSet({
-    forbidden: [
-      {
-        from: {
-          path: "^([^/]+)/.+",
-        },
-        to: {
-          circular: true,
-          viaOnly: { path: "^($1)/ab\\.js$" },
-        },
-      },
-    ],
-  });
-
-  it("a => ba => bb => bc => a doesn't get flagged when the cycle doesn't go via the viaOnly", () => {
+  it("does not flag when all of the cycle is inside the group-matched viaNot", () => {
     deepEqual(
       validate.dependency(
-        lCycleViaRuleSet,
-        { source: "tmp/a.js" },
+        parseRuleSet(lCycleButNotViaGroupMatchRuleSet),
+        { source: "src/module-a/a.js" },
         {
-          resolved: "tmp/ba.js",
+          resolved: "src/module-a/aa.js",
           circular: true,
-          cycle: ["tmp/ba.js", "tmp/bb.js", "tmp/bc.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
+          cycle: [
+            "src/module-a/aa.js",
+            "src/module-a/ab.js",
+            "src/module-a/ac.js",
+            "src/module-a/a.js",
+          ].map(stringToCycleEntry),
         },
       ),
       {
@@ -393,99 +287,36 @@ describe("[I] validate/index dependency - cycle viaOnly - with group matching", 
     );
   });
 
-  it("a => aa => ab => ac => does not get flagged when only some of them are not in the viaOnly", () => {
-    deepEqual(
-      validate.dependency(
-        lCycleViaRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: ["tmp/aa.js", "tmp/ab.js", "tmp/ac.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
-        },
-      ),
-      {
-        valid: true,
-      },
-    );
-  });
-
-  it("a => ab a gets flagged becaue all of the via's in the cycle are in the viaOnly", () => {
-    deepEqual(
-      validate.dependency(
-        lCycleViaRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: ["tmp/ab.js", "tmp/a.js"].map(stringToCycleEntry),
-        },
-      ),
-      {
-        valid: true,
-      },
-    );
-  });
-
-  it("a => aa => ab => ac => a get flagged when all of them are in a viaOnly", () => {
-    const lRuleSet = parseRuleSet({
+  it("does not flag when all of the cycle is inside the group-matched viaSomeNot that's represented as an array", () => {
+    const lRuleSet = {
       forbidden: [
         {
-          from: {},
+          name: "no-circular-dependency-of-modules",
+          from: { path: "^src/([^/]+)/.+" },
           to: {
+            viaSomeNot: "something|^src/$1/.+|somethingelse",
             circular: true,
-            viaOnly: "^tmp/[^.]+\\.js$",
           },
         },
       ],
-    });
+    };
     deepEqual(
       validate.dependency(
-        lRuleSet,
-        { source: "tmp/a.js" },
+        parseRuleSet(lRuleSet),
+        { source: "src/module-a/a.js" },
         {
-          resolved: "tmp/aa.js",
+          resolved: "src/module-a/aa.js",
           circular: true,
-          cycle: ["tmp/aa.js", "tmp/ab.js", "tmp/ac.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
+          cycle: [
+            "src/module-a/aa.js",
+            "src/module-a/ab.js",
+            "src/module-a/ac.js",
+            "src/module-a/a.js",
+          ].map(stringToCycleEntry),
         },
       ),
       {
-        valid: false,
-        rules: [{ name: "unnamed", severity: "warn" }],
-      },
-    );
-  });
-  it("a => aa => ab => ac => a get flagged when all of them are in a viaOnly presented as an array", () => {
-    const lRuleSet = parseRuleSet({
-      forbidden: [
-        {
-          from: {},
-          to: {
-            circular: true,
-            viaOnly: { path: "somethingelse|^tmp/[^.]+\\.js$" },
-          },
-        },
-      ],
-    });
-    deepEqual(
-      validate.dependency(
-        lRuleSet,
-        { source: "tmp/a.js" },
-        {
-          resolved: "tmp/aa.js",
-          circular: true,
-          cycle: ["tmp/aa.js", "tmp/ab.js", "tmp/ac.js", "tmp/a.js"].map(
-            stringToCycleEntry,
-          ),
-        },
-      ),
-      {
-        valid: false,
-        rules: [{ name: "unnamed", severity: "warn" }],
+        valid: true,
       },
     );
   });
